Implements inline mesh generation requested in issue #42
This commit adds the following three public methods to the MeshInfo class:
- meshRectangle()
- meshAnnulus() [for partial annuli]
- meshAnnulus() [for a full/closed annulus]
The implementation of the methods is put into the new files
- MeshGenRectangle.cpp
- MeshGenAnnulus.cpp
- MeshGenFromGmshFile.cpp
where the latter contains the source code for MeshInfo::fromGmshFile().
So MeshInfo.hpp/.cpp only contain source code for generic methods that
are useful for all variants.

The commit adds a new auxilliary method MeshInfo::deriveEdges(). This
allows to generate the edges of the mesh automatically, once the faces
and vertices have been constructed. It extracts the information from the
faces and sets the DoFType of an edge based on that of its two vertices.

The doxygen documentation of the class includes a description of the
new methods and what different flavours of mesh generation are supported.
For this the commit adds:
- figures of sample meshes in @tinyhhg_SOURCE_DIR@/doc/ExtraPics
- @tinyhhg_SOURCE_DIR@/doc/ExtraPics to Doxygen's IMAGE_PATH.

A new app called show_mesh is also added. It demonstrates the usage of
the new inline mesh generators and allows to output the resulting meshes
as VTU files for visualisation. It can also be used to import, export and
then visualise a Gmsh file.

**NOTE:**
The inline mesh generators currently set all vertex and edge primitives on
the domain boundary to DoFType DirichletBoundary and those inside the
domain, and of course all face primitives, to DoFType Inner.