Adds a script for auto-generation of operators for different elements
The commit adds a python script that allows generation of operators with
FFC from a form description that is independent of the actual element type.
Thus, we can auto-generator operators that are non-mixed from a single
UFL source file.

The script places the operator header files in the source tree and also
fixes the <ufc.h> header include path.