Fixes issue #63
We can have edges connecting vertices on different parts of the domain
boundary while lying in the interior. These were incorrectly flagged as
DirichletBoundary by the general MeshInfo::deriveEdges() method.

This commit removes this method and replaces it by two specialised
methods
- deriveEdgesForRectangles()
- deriveEdgesForFullAnnulus()
which are employed by the corresponding mesh generation routines and
placed in the respective source files.