Resolves issue #60
The commit resolves issue #60, i.e. it enhances our python script to also
generate the C++ header files for operators with mixed finite element pairs
automatically from generic UFC form files.

The changes are rather unintrusive. We now only distinguish in the script
and the form files between elementTest and elementTrial for the test and
trial spaces. The name and order for both elements is given in the elemList
of the script.

Due to the rather irregular placing of generated C++ files for mixed forms
in the source tree, we now need to specify their placement in the outDir
list.