Implemented 3D stencil assembly in P1ConstantOperator.
To decide on the "dimension" of the stencil (required during the
calculation of the stiffness matrix), the operator checks if
there are any cell-primitives in the global domain.

This also means that assembling a 2D operator on a 3D mesh
(and the other way round) will result in undefined behavior
(could segfault, could not...).

Also checking for DoFType of cell-primitives during apply.