Implements Jacobi, GS and SOR smoothing for P1 elementwise operator.
The commit implements smoothing of Jacobi, Gauss-Seidel and SOR type
for the P1ElementwiseOperator. All three smoothers are making use of
a single point_smooth_elementwise() function implemented for the vertex,
edge and face primitives.

Commit also adds a test that compares the result of a single smoothing
step using the P1ElementwiseOperator to that of the P1Operator.