First step towards more flexible boundary conditions (WIP).
Introducing a mesh boundary flag in the MeshInfo that shall in the future replace the dof type flag.
The mesh boundary flag is an integer which is assigned to each primitive of the mesh info.
The flag is set either in a mesh file or by a mesh generator.

In later steps, boundary conditions are assigned to the flags to allow grouping multiple primitives and
assigning them a BC.