Fixes issue #68 and extends functionality.
The commit implements getMaxValue() for the VertexDofMacroEdge and
VertexDofMacroVertex. These new functions are in turn used inside
VertexDoFFunction::getMaxValue(). This resolves the issue. The initial
halo updates in VertexDoFFunction::getMaxValue() are removed.

In a similar fashion two other auxiliary methods are implemented for
VertexDofFuntion: getMinValue() and getMaxMagnitude().

The test was renamed and extended.