Removed DoFType member from all primitives and refactored related code.
This commit also includes a minor other change:
The members of the Face class that specify neighboring primitives
that are located at the boundary have been removed.
Instead, the check can be performed through the SetupPrimitiveStorage.
This should be more general since it is implemented for all types
of primitives. Mainly affects the blending code.