Starts introduction of FaceDoFIndexing
As part of the refactoring (see issues #65, #75, etc) this commit
starts introduction FaceDoFindexing. The two files
- dgfunctionspace/DGFaceIndex.hpp
- bubblefunctionspace/BubbleFaceIndex.hpp
get deleted. Their functionality is now to be found inside
- facedofspace/FaceDoFIndexing.hpp
Note that we now have hhg::facedof::macroface as new namespace for
this stuff and that some function names have been shortened.

The commit updates the related tests which run through on my machine.