Completes introduction of FaceDoFIndexing
As part of the refactoring (see issues #65, #75, etc) this commit
completes the introduction FaceDoFindexing. The two files
- dgfunctionspace/DGEdgeIndex.hpp
- bubblefunctionspace/BubbleEdgeIndex.hpp
get deleted. Their functionality is now to be found inside
- facedofspace/FaceDoFIndexing.hpp
Note that we now have hhg::facedof::macroedge as new namespace for
this stuff and that some function names have been shortened.

I relocated the two implementations of DGEdge::printFunctionMemory()
to DGEdge.hpp as I did not see how these relate to indexing.

The commit updates the related tests which run through on my machine.