Refactors dot() method to avoid multiple reduce operations
This commit addresses issue #72. As discussed there it splits the
dot() method for computing an inner product into a dotLocal() and
a dotGlobal() method.

dotLocal() computes the inner product with respect to the primitives
owned by the local process. dotGlobal() builds on dotLocal() and performs
the reduce operation.

Low level functions:
-------------------

 The following functions now instead of dot() offer dotLocal():

 - EdgeDoFFunction
 - VertexDoFFunction

FE level functions:
------------------

 The following functions now instead of dot() offer both
 dotLocal() and dotGlobaLocal():

 - VertexDoFFunction (as there is no separate P1Function)
 - P2Function

Composite level functions:
-------------------------

 The following functions now instead of dot() offer dotGlobal():

 - P1StokesFunction
 - P2P1TaylorHoodFunction

Tests and apps are updated accordingly.