Added constructor to function base class that flags it as dummy.
The main purpose of this dummy feature is to be able to invoke all
methods on a function without consuming any compute time and to
be able to not allocate any data in the first place.

This way we can set the 3rd dim components (e.g. velocity in z-direction)
of composite functions to a dummy function during runtime (i.e. depending on the mesh).