Performs further code streamlining and adds additional max and min tests
As getMaxValue() and getMinValue() are now implemented also for DGFunction
and P2Function, we can run all test through the auxilliary function. This
makes the code much cleaner.

Extended test to check for maximal and minimal values, too, where this
had not been done, yet.

Changed combo test, so that the minimal value is -5 and thus the maximal
magnitude is +5.