BugFix: PETSc assembly and apply for macro-edge edgedof-to-vertexdof was broken.
The problem appeared due to wrong widths given to the basis conversion function.
For XYZ edges, the width is lower than for other edge types. This
issue does not appear for edge-to-edge because no XYZ type edges are touched.

Also this was already fixed for macro-faces but forgot to apply the same fix to macro-edges.

The new (PETSc-)symmetry test for the P2 Laplacian catches this issue.