Revives PolarLaplacian app
The old implementation of the PolarLaplacian app was based on the
P1ElementwiseOperator which we no longer support. Thus, the app was
deprecated.

This commit now revives the app as a test and demonstration of two
different approaches. Remember the app solves a Laplace problem in
polar coordinates on an annulus with Dirichlet BCs at the outer and
inner perimeter of the annulus and at an artifical inner boundary
at phi = 0 degree (no periodic BCs at the moment).

(1) One approach by which the app can solve the problem is by
    using the new PolarCoords geometry map in combination with
    the P1BlendingLaplaceOperator.

(2) The other alternative is to use the P1VariableOperator in
    combination with a P1FenicsForm which in turn employs the
    p1_polar_laplacian_form.

The approach can be selected via the "method" steering parameter.