Implements inline mesh generation for rectangular cuboids.
The commit implements a new function MeshInfo::meshCuboid() that
allows inline generation of a tetrahedral mesh for a rectangular
cuboid.

A demo using the new mesher is added to the show_mesh app.

Note that the current limitation is that the cuboid is axi-aligned.
However, this only affects computation of the vertex coordinates and
could easily be changed.