Truly eliminates Dirichlet DoFs from Matlab matrix
The commit adds code to hhg::exportOperator() that will append a vector
with the global indices of those DoFs that are fixed by Dirichlet
boundary conditions.

For this the variant of applyDirichletBCSymmetrically() that works only
on the matrix now returns this information to its caller.

Additionally Matlab code is appended to the script which truly removes
the Dirichlet DoFs from the sparse matrix.

The script also becomes a little bit more talkative.