Splitting the edge DoF SOR into multiple sweeps: one per edge type (generated).
This has severe performance advantages.
Due to the memory layout of the edge DoFs and the coincidence that there are
never two (or more) edges per type in one element, this split is a natural
coloring of the edge DoFs.
Therefore each of the split-kernels can be vectorized!
However it is not clear how the smoothing property suffers from this splitting.
In first tests it is marginally worse - but the performance gain is huge.