Adds new integrate() methods for 3D to P2FenicsForm.
The commit adds two new integrate() methods to the P2FenicsForm class.
These allow to request a single entry of the local element matrix, or
several selected entries from a single row.

The dof indexing is abstracted aways, i.e. the FEniCS indexing of the
local dofs is not explicitely used in the interface. Instead the
desired dofs (for row and column indices) are specified by pairs of
vertex indices using the P2Form::dofPosByVertexPair3D data type.