Re-adds implementation of old school integrate() for 3D
In a previous commit I had deleted my implementation for
void integrate( const std::array< Point3D, 4 >& coords, Point4D& out )
thinking this would no longer be required as we have the two new
integrate() methods for 3D.

However, I had to notice that the vertexToVertex sub-operator of our
P2ContantOperator is actually a P1ConstantOperator which employs a
P2form. Uffz.

As I currently have no intention on also touching the P1 operators,
this commit re-adds the implementation for the "old school" integrate()
for 3D and vertex-to-vertex couplings to P2FenicsForm.