Experimentally uses different integrate() routine in P2Elements3D.
This commit experimentally uses the integrate() member function of
P2FenicsForm that returns no a single but a selection of entries of
the local element matrix.

The routine is called in calculateVertexToEdgeStencilInMacroCell().
In order to get things working I added pseudo-implementations of this
special integrate() function to the P1ToP2FenicsForm and the
P2ToP1FenicsForm.