Modified Uzawa relaxation for the stable P2-P1 Stokes discretization.
The motivation is to get rid of a tmp variable in the Uzawa smoother
(this is necessary for extreme-scale scenarios as another tmp variable
may cost tens of TB of memory).

Instead of under-relaxing on the zero vector and adding to the solution,
we only multiply with the inverse of the diagonal of the PSPG operator.

This way we lose the contribution from the off-diagonals during the SOR.
However, this variant seems to perform quite well (even better than
before) in first tests.

Also it is now more similar to the variant proposed in Gaspar 2014.