Adds two algorithms for estimating the spectrum of operators
The commit adds two free functions implementing algorithms for
estimating the spectral radius or the extremal eigenvalues of
an operator. The algorithms are
- power iteration
- Lanczos approach based on conjugate gradients
For the second approach the CGSolver class now offers a method
to compute the tridiagonal matrix associated with the Lanczos method.

The commit also adds a test for the new algorithms which works on
the P1Laplacian on a regular mesh.