Adds relax parameter to P2ConstantOperator::smooth_jac()
So far the smooth_jac() method did not allow to set the relaxation
parameter. Instead deep down in P2MacroFace::smoothJacobiEdgeDoF and
P2MacroFace::smoothJacobiVertexDoF an optional factor of 2/3 was used.

Updated calls to smooth_jac() to use 2/3 or an appropriate value.

Also ran clang-format on the files of this commit, which resulted in
several formatting changes.