Adds check to high-level apply() methods that src != dst
In order to avoids problems with aliasing the commit adds a check for
debug mode that asserts that the addresses of src and dst are different.
Fixes this problem in the P2JacobiGSSmoothTest. There it did not cause,
problems, yet, as the strong rhs is zero for the current choice of
polynomial coefficients.