Adds a new P2ElementwiseOperator class.
This commit adds a new operator type that, like the old P1ElementwiseOperator
that we had in the past, is based on a different paradigm than the other
operators. Instead of working stencil-based it operates in an elementwise
fashion, i.e. we locally apply the element matrices to element-local vectors.

For this the P2FenicsForm class has a new integrateAll() method that
hands back the complete local element matrix.

The P2ConstantMassOperatorTest is changed to a P2MassOperatorTest as it
now also, successfully, tests the P2ElementwiseMassOperator.

Note: This is still work in progress; there are some issues with the
additive communication (see issue #111 in this context) and we can only
do 2D so far.