Adds a test for new indexing functionality w.r.t. micro-cells
This commit adds a minimal test for iterating over all micro-cells in
a macro-cell and getting the indi9ces of vertex and edge dofs associated
with a micro-cell.

The test is not very sophisticated. So it will probably not be sufficient
to assure correctness, but at least should help avoiding regressions.