Adds setToZero() method to scalar function classes
We add a setToZero() method to the following functions
- VertexDoFFunction (aka P1Function)
- EdgeDoFFunction
- P2Function
The method differs from calling interpolate() with a zero constant
in the sense that setToZero() will also set all DoFs in the halos of
the primitives to zero, while interpolate() only touches interior
DoFs. The new method is currently used by the P2ElementwiseOpertor.