Comments out extern template class statements
The commit comments out the following two statements
   extern template class VertexDoFFunction< double >;
   extern template class EdgeDoFFunction< double >;
in the corresponding header files of the two classes. For some
reason that is currently totally obscure to me I get errors of
the following form when the two statements are active and the
cpp files of the classes are compiled:

VertexDoFFunction.cpp:1765:95: error: specialization of ‘void hyteg::vertexdof::VertexDoFFunction<VType>::invertElementwise(walberla::uint_t, hyteg::DoFType) const [with ValueType = double; walberla::uint_t = long unsigned int]’ after instantiation
 void VertexDoFFunction< real_t >::invertElementwise( const uint_t level, const DoFType flag ) const

This happens with GCC 8.3. The error seems strange to me as my
understanding of the statements is that they suppress instantiation.