Worked on P1 macro-face evaluate:
The issue here was that in some cases, if the passed coordinate is
located right on the edge of the macro-face, this function would, due to
floating point arithmetic, determine invalid indices and then perform
invalid memory accesses.

The function now "projects" the given coordinates to the passed
macro-face to make sure that valid DoFs are returned.

A prerequisite is therefore that the passed coordinate _is_ located
in the macro-face (which it should be anyway). However, now we should
not get invalid memory accesses anymore and the function can be securely
called.