Implements matrix setup for elementwise operators
This commit implements the functionality required to assemble the matrix
associated with a P[12]ElementwiseOperator as a sparse PETScMatrix.

We also have a test that compares the matrix we get for the elementwise
operator to that of the corresponding P[12]ConstantOperator. This is done
for 2D and 3D with the operators for the Laplace and Mass forms.