Adds new HyTeG mass forms and a test comparing FEniCS and HyTeG
This commit adds a P2MassForm (with blending) for 2D and 3D settings
and a P1MassForm3D (with blending). These forms are placed in a separate
sub-directory as they are not fully automatically generated like the
existing forms in form_hyteg_base. They provide integrateAll() methods to
allow them to work together with the elementwise operators. We also
provide some standard quadrature rules in a new file.

Together with the forms there is also a new test that compares some of
the HyTeG forms to those generate with the FFC. Here we make sure that
if we do not use a geometryMap that the HyTeG forms compute a local
element matrix comparable to the one we get from FEniCS.