Adds first partially working version of IcosahedralShellMap
This commit adds a new geometry map to be used in combination with
icosahedral meshes for a thick spherical shell. Currently it allows
to map nodes on refined mesh levels to radial layers, i.e. the evalF()
method is already working. The 'show_mesh' app was extended to visualise
the results.

In order to make the map usable with forms we still need to implement
the computation of the Jacobian matrix.