Adds evaluation of inverse mapping to AnnulusMap and IdentityMap classes.
The new method evalFinv() allows to map back from the physical domain to
the computational domain. Currently it is only sensibly implemented for
the AnnulusMap and IdentityMap classes. Calling it on other blending types
will abort via the default implementation in the base class.

Results for the implementation in the AnnulusMap class look correct, see
the new AnnulusMapTest.