Adds draft for refactoring mesh generation for a full annulus
I suggest to implement a method

  meshAnnulus( const real_t rmin, const real_t rmax,
               const meshFlavour flavour, uint_t nTan, uint_t nRad );

that as the version for a partial annulus is based on meshing a rectangle
in polar coordinates. We only need to remove the duplicate nodes (with
phi = 0 or 2pi), fix the connectivity and mark the egdes with phi = 0
as interior.

Once the new method is finished, we can check it against

  meshAnnulus( const real_t rmin, const real_t rmax,
               uint_t nTan, uint_t nRad );

which directly meshes the full annulus for flavor=CRISSCROSS in cartesian
coordinates. Once the new method works, we could delete the old version
and the auxilliary deriveEdgesForFullAnnulus() method.