First major refactoring of UzawaSmoother.
This adds a new mandatory parameter to the Uzawa smoother that specifies
the smoother for the A block.

Until now, this was always a (symmetric or forward) Gauss-Seidel smoother.
However, with the new elementwise operators and also in general, we want
to apply other smoothers for the velocity part (e.g. Jacobi, Chebyshev-type).

Also this decouples the internal A-block smoother even more as a first step towards
smoothing on the epsilon-operator.

Also introduces a new solver class "StokesVelocityBlockBlockDiagonalPreconditioner"
which is the usual choice to be passed to the UzawaSmoother for constant coefficient
Stokes systems. This preconditioner can be passed any Solver instance that smoothes/solves
a scalar operator (for Stokes this is the Laplacian obviously). This Solver is applied
to each of the three block diagonals.