Introduced solver templates (!= c++ templates).
To minimize boilerplate code especially for simple
solvers in tests or apps where solver performance
is not important, we can in this directory define
functions that return solver instances for common
use cases.

As a first example, a simple configuration of the
Stokes gmg solver with inexact-Uzawa relaxation has
been added and directly integrated in two example test
cases.

The initial idea here is not to provide every possible
configuration for every possible application as a template
but to simplify implementation of solvers for new users
and to reduce code duplication for the majority of cases.
As often, the employed solvers are almost identical.