Replaces AffineMap class and completes GeometryMap::deserialize()
The commit replaces the AffineMap class by a more general version
AffineMap2D. The new one allows to describe the mapping via a 2x2
matrix and a translation vector.

The commit also adds new constructors to the AnnulusMap and
IcosahedralShellMap classes. These were needed to complete the
implementation of GeometryMap::deserialize().

We use the new AffineMap2D for additional blending tests in the
MassOperatorTest setting.