Corrected behavior of preconditioners in CG and MINRES solvers.
Until now, some "Solvers" (mainly the preconditioners) behaved
differently than the Solver interface suggests. The RHS
of the solver interface (b) was explicitly written to in
Solvers that shall be used as preconditioners.

This behavior was now corrected in the "preconditioners" as well
as in the available Krylov solvers (MINRES and CG). Only the solution
variable is now updated and intended to be updated by the implementation.