Completes implementation of P2ToP1FormDiv and P2FormDiv classes
This commit completes the implementation of two new HyTeG forms that
compute the weak divergence for a P2-P1 and a P2-P2 pair. We now also
take the blending, i.e. the form's GeometryMap into account.

Testing this is tedious for general maps. However, we can relatively
simply test at least a distortion of the domain into the x- and y-direction
plus a translation using AffineMap2D. As in this case we know how the
element matrices should scale w.r.t. the unmapped case.

The commit extends HytegVsFenicsFormTest to allow for such kind of tests
and adds them for some 2D forms.