Using split communicator in PETScLU solver.
All processes that own primitives perform the solve() call on the assembled problem.
All remaining processes solve an empty problem.

Intended for agglomeration: by simply re-distributing the primitives to a subset of processes,
the PETSc solver will solve the global problem only on that subset, the remaining processes
do not know about this and solve an empty problem. The solve() function may actually only
be invoked on non-empty processes then.