Changes testing strategy for affine maps in HytegVsFenicsFormTest()
Instead of scaling the element matrix we get back from FEniCS we now can
also apply the affine mapping used in the blending tests to our element
and feed the mapped element to FEniCS. This works as we use an affine map.

The advantage then is that we can apply this also from more challenging
mappings like rotate + scale + shift (as we now do for the 2D case) and
it will work with any form.

The calls to compareScaled() were replaced by calls to compareWithAffineMap().