Implemented PETSc assembly for the strong free-slip wrapper in 2D (T.-H. only).
First the individual matrices (Stokes, normal projection) are assembled,
then a PETSc MatMatMult() is performed and the result is stored.

The resulting operator is, as the matrix-free operator, not symmetric.
However, the PETSc MINRES solver seems to converge anyway.

Added a small test for the assembly (checking if HyTeG and PETSc
operator application produce same results).

Extended free-slip test and app.