Resolves stability issue with P2Form_laplace in 3D.
In the code generator we factored out the determinants from the inverse
Jacobi matrices. And add them again separately later. Now the only
divisions in the generated code are related to factoring the determinants
back in.

The HyTeGVsFenicsFormTest now runs through with the vertex coordinates
that caused the problem in issue #133.