Updates P2P1ElementwiseBlendingStokesOperator
An operator needs to typedef a BlockPreconditioner_T to be usable in
combination with the StrongFreeSlipOperator.

As a temporary fix we use the P2P1TaylorHoodStokesBlockPreconditioner
for the P2P1ElementwiseBlendingStokesOperator. This is likely to be
suboptimal as the preconditioner does not include blending, but must
do for the moment.