Implements a P1RowSumForm
Commit adds a P1RowSumForm which is a child of the P1Form. Same as with
the P2RowSumForm. The P1RowSumForm lumps the weights of a P1Form onto
the diagonal.

In order to make it usable with P1HyTeGForms the commit also makes the
latter a child of the P1Form. Something we did not do so far, because
our form stuff is not too consistent currently w.r.t. all the different
integrate() variants and the fact that we have the generated HyTeG forms
which so far only work for 2D and the "manual" HyTeG forms that for P1
elements are partially 3D only, while the P2 HyTeG forms cover 2D and 3D.

The RowSumFormTests now include the P1RowSumForm using the same
P1FenicsForms as the test for P2.

Note that the P1RowSumForm is not really needed for the P1FenicsForms
as the P1ConstantOperator inherently implements lumping by itself.