Implements a DiagonalNonConstantOperator class
The new class constitutes an operator that is, in a matrix sense, diagonal
and where the diagonal values may change from DoF to DoF. An example would
be a lumped mass matrix on a blended domain.

The operator wraps itself around an ElementwiseOperator and uses the latter
to assemble and obtain the diagonal values. The DiagonalNonConstantOperator
is intended to work together with RowSumForms, but will work for any other
form, too, by just disregarding the non-diagonal matrix entries.

While currently combined with an underlying ElementwiseOperator the
DiagonalNonConstantOperator could also be combined with any other operator,
as long as that provides access to the diagonal matrix entries in the form
of a HyTeG function.