Fixes signature issues in P[12]VectorFunction::interpolate()
For the case that we pass an std::vector of std::functions the
interface of the functions was missing an & in the Point3D argument.
Also we add const qualifier to the std::vector.

The P2VectorFunctionTest now uses this variant for testing and
source code of VTKOutputTest needed adaptation.