Adds a new faster variant of the 3D blended Laplacian form
In the new variant P2Form_laplacePimped3D the ordering of "loops" is
re-arranged so that the outermost loop is over the integration points.
We then precomputed for the mapped point all shape function derivatives.
This reduces work significantly and leads to faster execution in the
P2Benchmark benchmark.

Added the new form to the HyTeGVsFenicsFormTest to be sure that it
computes the "real thing".