Adds draft of a generic VectorToVectorOperator
The operator maps one VectorFunction onto another one. Both can potentially
be of different type. The operator assumes that the function spaces are
simply extensions of scalar spaces that that we have a VectorComponentType
available.

Currently the class only implements a constructor, an apply() method and
a setSubOperator() method. The latter is used in the factory approach,
see below.

We also implement
* a generic P1VectorOperator(); this one is initially empty and can be
  filled using setSubOperator(); which we do in the sample factory
   implementation generateP1VectorLaplaceOperator()
* as an alternative we also derive by inheritance a
  P2ConstantVectorLaplaceOperator

Both approaches are used in the vecOpDemo app to apply the 2D Laplacian
to two vector valued functions interpolated into the P1 and P2 spaces.

Note that we needed to add a virtual apply() function in the Operator
class for this to work.