Starts replacing P[12]VectorFunction be their new AltKind implemtations
Lots of replacing .u, .v, .w by [0], [1], [2]; some code reduction by
combining scalar interpolate() and assign() calls.

NOTE:

We have a design problem with assign() and add(). While the compiler can
implicitely upcast a P2VectorFunction< vType > to a
CFSVectorFunction< P2Function< vType > > in an initialiser list, it cannot
do the same with the entries of an std::vector. This leads to problems
e.g. in some other classes.

Probably we can solve this be switching to true CRTP using
CFSVectorFunction< P2VectorFunction< vType > >.

Also we might run into problems with the enumerator stuff. But we will
see.