Updates w.r.t. forms for Epsilon operator
Up to this commit the FEniCS forms for the Epsilon operator are not
usable in 3D, because the corresponding code in ufc_traits.hpp was
missing. Commit fixes this.

We also add comparision of the Epsilon forms generated via HFG and
FFC to the HytegVsFenicsFormTest for P1, P2 and 2D, 3D w/o blending.