Adds new enumeration type functionTraits::FunctionKind
The new enumeration allows to implement a query method
GenericFunction::getFunctionKind() that can be used to obtain
information on the underlying (wrapped) function. The previous
type traits do not help here, because we cannot return a datatype
(and working with string from getTypeName() feels a little clumsy.

We use the FunctionKind value inside VTKOutput::add() in case of a
GenericFunction argument to find out the single type cast we should
attempt.