Adds three classes for P2 Epsilon Operators
The new classes are children of the VectorToVectorOperator and called
1. P2ConstantEpsilonOperator
2. P2ElementwiseAffineEpsilonOperator
3. P2ElementwiseBlendingEpsilonOperator

Operator 1 uses the FEniCS forms p2_(tet_)stokes_epsilon(_tet) while
the other use corresponding HyTeG forms. These are the var variants
and, thus, need a callback function for the variable vicosity. This
is to be passed to the constructor of the operators. They will use
it to generate objects for the sub-forms and pass them to the
constructors of their Elementwise sub-operators.

Additionally commit implements a P2ConstantFullViscousOperator.
The latter is based on FEniCS forms. So it does support neither
blending nor variable viscosity.

Adds a ViscousOperatorsTest that checks generation of objects of the
four new operator types for a 2D and a 3D setting.