Implements integrate() for P1FormHyTeG
The new HFG forms do not sport an integrate() method. Although they have
an integrateRow0() method which does the same thing. However, for replacing
some of the deprecated old forms integrate() is required. Since the manual
HyTeG forms do no offer either integrate() or integrateRow0() this commit
does a transitional implementation based in integrateAll() which is of course
slow. Cleanup is definitely required.