Adds a PetscMatrixAssemblyTest
We already individually test for some of the operators that they can be
assembled into a PetscSparseMatrix. This new test does this explicitely
for a wide range of operators/operator families, i.e. we it does not test
all P1ConstantOperators e.g. but only one of them.

Note that operators for which an assembly does currently not work are
only reported, but not tested, so the test succeeds currently.

The commit places the test together with PetscTest.cpp into a new Petsc
sub-directory.