Starts disentangling of VTKOutput class
Commits places the following free functions in the new file VTKHelpers.hpp
- writeXMLHeader()
- writeXMLFooter()
- writePieceHeader()
- writePieceFooter()
They now belong to the namespace vtk.

We also convert VTKOutput::openDataElement() into a free function, placed
in the same file and namespace as above.

Additionally we remove the class enum VTKOutput::VTK_DATA_FORMAT and replace
it by vtk::DataFormat.

The private methods
- writeP1()
- writeVertexDoFData()
- writeP1VectorFunctionData()
are now members of a new VTKP1Writer friend class.

The private methods
- writeP2()
- writeSingleP2Function()
- writeSingleP2VectorFunction()
are now members of a new VTKP2Writer friend class.
The private method
- writeDGDoF()
is now a member of a new VTKDGDoFWriter friend class.

Finally we modify the VTKOutputTest so that some data are actually written
in base64 encoded binary